[ISSUE #12017] Fix errors
* Add ApiType annotation

* Transfer the updateCluster method

* Fix Compilation Errors
