+++ console/src/main/java/com/alibaba/nacos/console/controller/v3/core/ConsoleClusterController.java
@@ -17,36 +17,24 @@
 
 package com.alibaba.nacos.console.controller.v3.core;
 
-import com.alibaba.nacos.api.common.Constants;
 import com.alibaba.nacos.api.model.v2.Result;
-import com.alibaba.nacos.api.naming.CommonParams;
-import com.alibaba.nacos.api.naming.pojo.healthcheck.AbstractHealthChecker;
-import com.alibaba.nacos.api.naming.pojo.healthcheck.HealthCheckerFactory;
 import com.alibaba.nacos.auth.annotation.Secured;
-import com.alibaba.nacos.common.utils.ConvertUtils;
-import com.alibaba.nacos.common.utils.NumberUtils;
-import com.alibaba.nacos.common.utils.StringUtils;
 import com.alibaba.nacos.config.server.paramcheck.ConfigDefaultHttpParamExtractor;
 import com.alibaba.nacos.console.proxy.core.ClusterProxy;
 import com.alibaba.nacos.core.cluster.Member;
 import com.alibaba.nacos.core.paramcheck.ExtractorManager;
 import com.alibaba.nacos.core.utils.Commons;
-import com.alibaba.nacos.core.utils.WebUtils;
-import com.alibaba.nacos.naming.core.v2.metadata.ClusterMetadata;
-import com.alibaba.nacos.naming.misc.UtilsAndCommons;
 import com.alibaba.nacos.plugin.auth.constant.ActionTypes;
 import com.alibaba.nacos.plugin.auth.constant.SignType;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.Collection;
 
 /**
- * .
+ * Controller for handling HTTP requests related to cluster operations.
  *
  * @author zhangyukun on:2024/8/16
  */
@@ -77,36 +65,5 @@ public class ConsoleClusterController {
     public Result<Collection<Member>> getNodeList(@RequestParam(value = "keyword", required = false) String ipKeyWord) {
         Collection<Member> result = clusterProxy.getNodeList(ipKeyWord);
         return Result.success(result);
-    }
-    
-    
-    /**
-     * Update cluster.
-     *
-     * @param request http request
-     * @return 'ok' if success
-     * @throws Exception if failed
-     */
-    @PutMapping
-    @Secured(action = ActionTypes.WRITE)
-    public Result<String> updateCluster(HttpServletRequest request) throws Exception {
-        final String namespaceId = WebUtils.optional(request, CommonParams.NAMESPACE_ID,
-                Constants.DEFAULT_NAMESPACE_ID);
-        final String clusterName = WebUtils.required(request, CommonParams.CLUSTER_NAME);
-        final String serviceName = WebUtils.required(request, CommonParams.SERVICE_NAME);
-        
-        ClusterMetadata clusterMetadata = new ClusterMetadata();
-        clusterMetadata.setHealthyCheckPort(NumberUtils.toInt(WebUtils.required(request, "checkPort")));
-        clusterMetadata.setUseInstancePortForCheck(
-                ConvertUtils.toBoolean(WebUtils.required(request, "useInstancePort4Check")));
-        AbstractHealthChecker healthChecker = HealthCheckerFactory.deserialize(
-                WebUtils.required(request, "healthChecker"));
-        clusterMetadata.setHealthChecker(healthChecker);
-        clusterMetadata.setHealthyCheckType(healthChecker.getType());
-        clusterMetadata.setExtendData(
-                UtilsAndCommons.parseMetadata(WebUtils.optional(request, "metadata", StringUtils.EMPTY)));
-        
-        clusterProxy.updateClusterMetadata(namespaceId, serviceName, clusterName, clusterMetadata);
-        return Result.success("ok");
     }
 }
